backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
opencv-python-headless==4.9.0.80
numpy==1.26.3
scipy==1.12.0
boto3==1.34.34
websockets==12.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.1
